+++ packages/orq-rc/src/orq_ai_sdk/agents.py
@@ -3,15 +3,6 @@
 from .basesdk import BaseSDK
 from orq_ai_sdk import models, utils
 from orq_ai_sdk._hooks import HookContext
-from orq_ai_sdk.models import (
-    createagentop as models_createagentop,
-    invokeagentop as models_invokeagentop,
-    listagenttasksop as models_listagenttasksop,
-    runagentop as models_runagentop,
-    streamagentop as models_streamagentop,
-    streamrunagentop as models_streamrunagentop,
-    updateagentop as models_updateagentop,
-)
 from orq_ai_sdk.types import BaseModel, OptionalNullable, UNSET
 from orq_ai_sdk.utils import eventstreaming, get_security_from_env
 from orq_ai_sdk.utils.unmarshal_json_response import unmarshal_json_response
@@ -217,10 +208,7 @@ class Agents(BaseSDK):
         self,
         *,
         request: Optional[
-            Union[
-                models_createagentop.CreateAgentRequestBody,
-                models_createagentop.CreateAgentRequestBodyTypedDict,
-            ]
+            Union[models.CreateAgentRequestBody, models.CreateAgentRequestBodyTypedDict]
         ] = None,
         retries: OptionalNullable[utils.RetryConfig] = UNSET,
         server_url: Optional[str] = None,
@@ -319,10 +307,7 @@ class Agents(BaseSDK):
         self,
         *,
         request: Optional[
-            Union[
-                models_createagentop.CreateAgentRequestBody,
-                models_createagentop.CreateAgentRequestBodyTypedDict,
-            ]
+            Union[models.CreateAgentRequestBody, models.CreateAgentRequestBodyTypedDict]
         ] = None,
         retries: OptionalNullable[utils.RetryConfig] = UNSET,
         server_url: Optional[str] = None,
@@ -988,35 +973,29 @@ class Agents(BaseSDK):
         instructions: Optional[str] = None,
         system_prompt: Optional[str] = None,
         model: Optional[
-            Union[
-                models_updateagentop.UpdateAgentModel,
-                models_updateagentop.UpdateAgentModelTypedDict,
-            ]
+            Union[models.UpdateAgentModel, models.UpdateAgentModelTypedDict]
         ] = None,
         fallback_models: Optional[
             Union[
-                List[models_updateagentop.UpdateAgentFallbackModels],
-                List[models_updateagentop.UpdateAgentFallbackModelsTypedDict],
+                List[models.UpdateAgentFallbackModels],
+                List[models.UpdateAgentFallbackModelsTypedDict],
             ]
         ] = None,
         settings: Optional[
-            Union[
-                models_updateagentop.UpdateAgentSettings,
-                models_updateagentop.UpdateAgentSettingsTypedDict,
-            ]
+            Union[models.UpdateAgentSettings, models.UpdateAgentSettingsTypedDict]
         ] = None,
         path: Optional[str] = None,
         memory_stores: Optional[List[str]] = None,
         knowledge_bases: Optional[
             Union[
-                List[models_updateagentop.UpdateAgentKnowledgeBases],
-                List[models_updateagentop.UpdateAgentKnowledgeBasesTypedDict],
+                List[models.UpdateAgentKnowledgeBases],
+                List[models.UpdateAgentKnowledgeBasesTypedDict],
             ]
         ] = None,
         team_of_agents: Optional[
             Union[
-                List[models_updateagentop.UpdateAgentTeamOfAgents],
-                List[models_updateagentop.UpdateAgentTeamOfAgentsTypedDict],
+                List[models.UpdateAgentTeamOfAgents],
+                List[models.UpdateAgentTeamOfAgentsTypedDict],
             ]
         ] = None,
         retries: OptionalNullable[utils.RetryConfig] = UNSET,
@@ -1165,35 +1144,29 @@ class Agents(BaseSDK):
         instructions: Optional[str] = None,
         system_prompt: Optional[str] = None,
         model: Optional[
-            Union[
-                models_updateagentop.UpdateAgentModel,
-                models_updateagentop.UpdateAgentModelTypedDict,
-            ]
+            Union[models.UpdateAgentModel, models.UpdateAgentModelTypedDict]
         ] = None,
         fallback_models: Optional[
             Union[
-                List[models_updateagentop.UpdateAgentFallbackModels],
-                List[models_updateagentop.UpdateAgentFallbackModelsTypedDict],
+                List[models.UpdateAgentFallbackModels],
+                List[models.UpdateAgentFallbackModelsTypedDict],
             ]
         ] = None,
         settings: Optional[
-            Union[
-                models_updateagentop.UpdateAgentSettings,
-                models_updateagentop.UpdateAgentSettingsTypedDict,
-            ]
+            Union[models.UpdateAgentSettings, models.UpdateAgentSettingsTypedDict]
         ] = None,
         path: Optional[str] = None,
         memory_stores: Optional[List[str]] = None,
         knowledge_bases: Optional[
             Union[
-                List[models_updateagentop.UpdateAgentKnowledgeBases],
-                List[models_updateagentop.UpdateAgentKnowledgeBasesTypedDict],
+                List[models.UpdateAgentKnowledgeBases],
+                List[models.UpdateAgentKnowledgeBasesTypedDict],
             ]
         ] = None,
         team_of_agents: Optional[
             Union[
-                List[models_updateagentop.UpdateAgentTeamOfAgents],
-                List[models_updateagentop.UpdateAgentTeamOfAgentsTypedDict],
+                List[models.UpdateAgentTeamOfAgents],
+                List[models.UpdateAgentTeamOfAgentsTypedDict],
             ]
         ] = None,
         retries: OptionalNullable[utils.RetryConfig] = UNSET,
@@ -1335,23 +1308,14 @@ class Agents(BaseSDK):
         self,
         *,
         key: str,
-        message: Union[
-            models_invokeagentop.Message, models_invokeagentop.MessageTypedDict
-        ],
+        message: Union[models.Message, models.MessageTypedDict],
         task_id: Optional[str] = None,
         variables: Optional[Dict[str, Any]] = None,
-        contact: Optional[
-            Union[models_invokeagentop.Contact, models_invokeagentop.ContactTypedDict]
-        ] = None,
+        contact: Optional[Union[models.Contact, models.ContactTypedDict]] = None,
         thread: Optional[
-            Union[
-                models_invokeagentop.InvokeAgentThread,
-                models_invokeagentop.InvokeAgentThreadTypedDict,
-            ]
+            Union[models.InvokeAgentThread, models.InvokeAgentThreadTypedDict]
         ] = None,
-        memory: Optional[
-            Union[models_invokeagentop.Memory, models_invokeagentop.MemoryTypedDict]
-        ] = None,
+        memory: Optional[Union[models.Memory, models.MemoryTypedDict]] = None,
         metadata: Optional[Dict[str, Any]] = None,
         retries: OptionalNullable[utils.RetryConfig] = UNSET,
         server_url: Optional[str] = None,
@@ -1466,23 +1430,14 @@ class Agents(BaseSDK):
         self,
         *,
         key: str,
-        message: Union[
-            models_invokeagentop.Message, models_invokeagentop.MessageTypedDict
-        ],
+        message: Union[models.Message, models.MessageTypedDict],
         task_id: Optional[str] = None,
         variables: Optional[Dict[str, Any]] = None,
-        contact: Optional[
-            Union[models_invokeagentop.Contact, models_invokeagentop.ContactTypedDict]
-        ] = None,
+        contact: Optional[Union[models.Contact, models.ContactTypedDict]] = None,
         thread: Optional[
-            Union[
-                models_invokeagentop.InvokeAgentThread,
-                models_invokeagentop.InvokeAgentThreadTypedDict,
-            ]
+            Union[models.InvokeAgentThread, models.InvokeAgentThreadTypedDict]
         ] = None,
-        memory: Optional[
-            Union[models_invokeagentop.Memory, models_invokeagentop.MemoryTypedDict]
-        ] = None,
+        memory: Optional[Union[models.Memory, models.MemoryTypedDict]] = None,
         metadata: Optional[Dict[str, Any]] = None,
         retries: OptionalNullable[utils.RetryConfig] = UNSET,
         server_url: Optional[str] = None,
@@ -1600,7 +1555,7 @@ class Agents(BaseSDK):
         limit: Optional[float] = 10,
         starting_after: Optional[str] = None,
         ending_before: Optional[str] = None,
-        status: Optional[models_listagenttasksop.Status] = None,
+        status: Optional[models.Status] = None,
         retries: OptionalNullable[utils.RetryConfig] = UNSET,
         server_url: Optional[str] = None,
         timeout_ms: Optional[int] = None,
@@ -1706,7 +1661,7 @@ class Agents(BaseSDK):
         limit: Optional[float] = 10,
         starting_after: Optional[str] = None,
         ending_before: Optional[str] = None,
-        status: Optional[models_listagenttasksop.Status] = None,
+        status: Optional[models.Status] = None,
         retries: OptionalNullable[utils.RetryConfig] = UNSET,
         server_url: Optional[str] = None,
         timeout_ms: Optional[int] = None,
@@ -1809,10 +1764,7 @@ class Agents(BaseSDK):
         self,
         *,
         request: Optional[
-            Union[
-                models_runagentop.RunAgentRequestBody,
-                models_runagentop.RunAgentRequestBodyTypedDict,
-            ]
+            Union[models.RunAgentRequestBody, models.RunAgentRequestBodyTypedDict]
         ] = None,
         retries: OptionalNullable[utils.RetryConfig] = UNSET,
         server_url: Optional[str] = None,
@@ -1905,10 +1857,7 @@ class Agents(BaseSDK):
         self,
         *,
         request: Optional[
-            Union[
-                models_runagentop.RunAgentRequestBody,
-                models_runagentop.RunAgentRequestBodyTypedDict,
-            ]
+            Union[models.RunAgentRequestBody, models.RunAgentRequestBodyTypedDict]
         ] = None,
         retries: OptionalNullable[utils.RetryConfig] = UNSET,
         server_url: Optional[str] = None,
@@ -2002,8 +1951,8 @@ class Agents(BaseSDK):
         *,
         request: Optional[
             Union[
-                models_streamrunagentop.StreamRunAgentRequestBody,
-                models_streamrunagentop.StreamRunAgentRequestBodyTypedDict,
+                models.StreamRunAgentRequestBody,
+                models.StreamRunAgentRequestBodyTypedDict,
             ]
         ] = None,
         retries: OptionalNullable[utils.RetryConfig] = UNSET,
@@ -2115,8 +2064,8 @@ class Agents(BaseSDK):
         *,
         request: Optional[
             Union[
-                models_streamrunagentop.StreamRunAgentRequestBody,
-                models_streamrunagentop.StreamRunAgentRequestBodyTypedDict,
+                models.StreamRunAgentRequestBody,
+                models.StreamRunAgentRequestBodyTypedDict,
             ]
         ] = None,
         retries: OptionalNullable[utils.RetryConfig] = UNSET,
@@ -2227,29 +2176,17 @@ class Agents(BaseSDK):
         self,
         *,
         key: str,
-        message: Union[
-            models_streamagentop.StreamAgentMessage,
-            models_streamagentop.StreamAgentMessageTypedDict,
-        ],
+        message: Union[models.StreamAgentMessage, models.StreamAgentMessageTypedDict],
         task_id: Optional[str] = None,
         variables: Optional[Dict[str, Any]] = None,
         contact: Optional[
-            Union[
-                models_streamagentop.StreamAgentContact,
-                models_streamagentop.StreamAgentContactTypedDict,
-            ]
+            Union[models.StreamAgentContact, models.StreamAgentContactTypedDict]
         ] = None,
         thread: Optional[
-            Union[
-                models_streamagentop.StreamAgentThread,
-                models_streamagentop.StreamAgentThreadTypedDict,
-            ]
+            Union[models.StreamAgentThread, models.StreamAgentThreadTypedDict]
         ] = None,
         memory: Optional[
-            Union[
-                models_streamagentop.StreamAgentMemory,
-                models_streamagentop.StreamAgentMemoryTypedDict,
-            ]
+            Union[models.StreamAgentMemory, models.StreamAgentMemoryTypedDict]
         ] = None,
         metadata: Optional[Dict[str, Any]] = None,
         stream_timeout_seconds: Optional[float] = None,
@@ -2385,29 +2322,17 @@ class Agents(BaseSDK):
         self,
         *,
         key: str,
-        message: Union[
-            models_streamagentop.StreamAgentMessage,
-            models_streamagentop.StreamAgentMessageTypedDict,
-        ],
+        message: Union[models.StreamAgentMessage, models.StreamAgentMessageTypedDict],
         task_id: Optional[str] = None,
         variables: Optional[Dict[str, Any]] = None,
         contact: Optional[
-            Union[
-                models_streamagentop.StreamAgentContact,
-                models_streamagentop.StreamAgentContactTypedDict,
-            ]
+            Union[models.StreamAgentContact, models.StreamAgentContactTypedDict]
         ] = None,
         thread: Optional[
-            Union[
-                models_streamagentop.StreamAgentThread,
-                models_streamagentop.StreamAgentThreadTypedDict,
-            ]
+            Union[models.StreamAgentThread, models.StreamAgentThreadTypedDict]
         ] = None,
         memory: Optional[
-            Union[
-                models_streamagentop.StreamAgentMemory,
-                models_streamagentop.StreamAgentMemoryTypedDict,
-            ]
+            Union[models.StreamAgentMemory, models.StreamAgentMemoryTypedDict]
         ] = None,
         metadata: Optional[Dict[str, Any]] = None,
         stream_timeout_seconds: Optional[float] = None,