+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasetop.py
@@ -93,5 +93,5 @@ class CreateDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-11-05T11:09:10.226Z")
+    updated: Optional[datetime] = parse_datetime("2025-11-05T15:06:44.874Z")
     r"""The date and time the resource was last updated"""