+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -230,7 +230,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K99V9YJG0B7P70CH5421JBEZ"
+        "01K9A8X0TXRPCGTAGDDWKRKA70"
     )
     r"""The unique identifier of the data source"""