+++ packages/orq-rc/src/orq_ai_sdk/models/filelistop.py
@@ -79,7 +79,7 @@ class FileListData(BaseModel):
     workspace_id: str
     r"""The id of the resource"""
 
-    created: Optional[datetime] = parse_datetime("2025-11-05T11:09:13.822Z")
+    created: Optional[datetime] = parse_datetime("2025-11-05T15:06:48.687Z")
     r"""The date and time the resource was created"""