+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K99V9YJGKR5T7RGZ415GECFD"
+        "01K9A8X0TV7B0A64Z1VCV0N8D3"
     )
     r"""The unique identifier of the data source"""